Extract cmdlog package and split git into files by concept
Separate command logging (cmdlog) from table rendering (display) to
eliminate the serverPort() duplication caused by the opencode->display
import cycle. Split git/git.go (463 lines, 4 concepts) into git.go,
classify.go, diff.go, and worktree.go so file names reflect the
architecture.

Import graph after: display is consumed only by main (presentation
layer), ssh/git/opencode import cmdlog (cross-cutting trace), no
cycles, no duplication.
