+++ pkg/cmdlog/cmdlog.go
@@ -1,4 +1,4 @@
-package display
+package cmdlog
 
 import (
 	"fmt"
