+++ pkg/opencode/server.go
@@ -8,7 +8,7 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/ellistarn/wt/pkg/display"
+	"github.com/ellistarn/wt/pkg/cmdlog"
 	"github.com/ellistarn/wt/pkg/ssh"
 )
 
@@ -44,7 +44,7 @@ func EnsureLocalServer() error {
 	}
 
 	cmd := exec.Command(binary, "serve", "--port", strconv.Itoa(port))
-	display.LogCmd(fmt.Sprintf("opencode serve --port %d", port))
+	cmdlog.LogCmd(fmt.Sprintf("opencode serve --port %d", port))
 	cmd.SysProcAttr = &syscall.SysProcAttr{Setsid: true}
 	devNull, err := os.OpenFile(os.DevNull, os.O_WRONLY, 0)
 	if err == nil {
@@ -84,7 +84,7 @@ func EnsureRemoteServer(host string) error {
 
 	port := ServerPort()
 	startCmd := fmt.Sprintf("$SHELL -ic 'nohup opencode serve --port %d </dev/null >/dev/null 2>&1 &'", port)
-	display.LogCmd(fmt.Sprintf("%s: opencode serve --port %d", host, port))
+	cmdlog.LogCmd(fmt.Sprintf("%s: opencode serve --port %d", host, port))
 	if _, err := ssh.Run(host, startCmd); err != nil {
 		// Race — someone else may have started it.
 		if healthProbe(tunnelURL) == nil {
